refactor: HotPlaceData 를 types 계층으로 이동
매퍼(utils/api)가 컴포넌트에서 UI 타입을 가져오면 하위 계층이 상위 계층에 묶여
카드를 옮기거나 쪼갤 때 유틸이 깨지고, 카드가 매퍼 헬퍼를 쓰면 순환 참조가 된다.
settingsMapper·planMapper 와 같은 방향(types/·api/types 에서 가져오기)으로 맞춘다.
